upload: remove useless permission
From https://developers.google.com/drive/api/v2/reference/permissions/insert#request-body

"The email address or domain name for the entity. This is used during inserts and is not populated
in responses. When making a drive.permissions.insert request, exactly one of the id or value fields
must be specified unless the permission type is anyone, in which case both id and value are ignored."

Since we set type: anyone, this leaves value useless

Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>